Properly invalidate non-local stores in global load/store opts.
In global load/store optimization, ensure that we remove non-local
stores from the store map when we hit instructions that can potentially
overwrite all or part of the stored value.

In the test case from this particular radar, the address being stored to
is also passed as an inout to a mutating method, and then loaded after
the mutation.

rdar://problem/20603219

Swift SVN r28053