[test] Add type annotations to Algorithm.swift.gyb.
This cuts the compile time from four minutes to two with a debug compiler.
However, two minutes is still a long time, even with quite a bit of AST and
SIL to verify, so it may still be worthwhile to figure out what's going on.
(The test itself takes another minute to run on my machine.)

No intended functionality change.

Swift SVN r28147