To preserve source level copies, is_unique must appear to decrement the refcount of its object.
Fixes <rdar://problem/20901275> Optimize retain/release across is_unique when the pointer cannot alias.

This is not a regression, but was something that adding SIL is_unique
is supposed to fix. However, that required making ARC optimization
more conservative. Unfortunately it is necessary for correctness. We
can improve the situation later with more advanced analysis:

<rdar://problem/20901275> Optimize retain/release across is_unique when the pointer cannot alias

Swift SVN r28438