Dependencies: add tests for protocol extensions.
I think this is all working fine: we already say that a type adopting a
protocol depends on that protocol's members, and that an extension of a
protocol can supply members. It's a pretty conservative mechanism since
it doesn't take the constraints into account, but that's okay for now.

rdar://problem/20476867

Swift SVN r28475