Handle try_apply in the diagnostic unreachable-code pass.
A somewhat more natural solution would be to replace the
normal successor with an edge to an unreachable block,
but that would require adding blocks during the iteration.

Add a small amount of SIL infrastructure for asking a
pred_iterator exactly which edge out of a terminator it
represents.

Swift SVN r28514