Fix specialization of protocols that return Self type.
In propagateConcreteTypeOfInitExistential we are specializing apply
instructions from a protocol type to a concrete type. However, we are not
handling Self return types (that are implemented as an argument that
is pointer to a stack allocation of the protocol type). This commit disables
the specialization of methods with the Self return type.

rdar://20868966

Swift SVN r28553