Avoid a crash-on-invalid by passing contextualType by reference to solveForExpression.
solveForExpression ends up storing the address of the contextual
type in the ConstraintSystem, so it's important that it live at
least as long as the system.  I don't actually know why the
ConstraintSystem has to store the address of a type, but it
does, so this is what we have to do.

Swift SVN r28593