ClosureSpecializer: Only specialize if the partial_apply and its user are in the same block
This is to workaround the issue in rdar://20883911 "Algorithm.swift.gyb segfaults
at runtime, if compiled with -O".

Closure specialization currently does not handle partial_apply arguments
correctly if the partial apply is not executed the same number of times as
the function call that takes it as argument.

I did not measure a noticeable effect on the PerfTestSuite at -O and -Onone.

rdar://20883911

Swift SVN r28597