Fix optimizeUnconditionalCheckedCastInst
It is not safe to dyn_cast the result to a SILInstrucion. Users of this function
expect a nullptr return to mean no change has happened.  If the operand was an
argument this cast would fail although we have removed the cast instruction and
changed the function.

rdar://20991580

Swift SVN r28684