SIL: Generate emitRawApply()'s @guaranteed cleanups with CleanupLocation
These used a RegularLocation because the mechanism is different from
the usual Cleanup stuff, which is tied to a specific scope.

Using a CleanupLocation prevents unreachable code diagnostics from
complaining if we call a @noreturn with guaranteed.

If we diagnose while emitting one of the cleanups, the diagnostic will
point to the end of the apply expression rather than the start, but I
don't think that matters.

Swift SVN r28690