SIL: propagate @noreturn while uncurrying function type, just like throws
When lowering a function type, SIL would lose @noreturn if the outermost
function was not @noreturn and the uncurry level was not 0.

This fixes a regression from r28306 "fix @noreturn and throws for curried
functions". Previously, we would incorrectly set @noreturn on each level
of a curried function. Once that was fixed, SIL function type lowering
broke as above.

Swift SVN r28691