SIL: look at lowered function type when diagnosing issues with returns
Checking the @noreturn property of the outermost function of the
FuncDecl's type is wrong, because the method body itself may have
an uncurried type. In particular, this meant that @noreturn
diagnostics were broken for methods.

Since we went through the pain of correctly computing the @noreturn
of the lowered type already, just use that instead.

Swift SVN r28692