Parser: Extend the "builder pattern" hack for properties and generic types.
Factor out the special case condition for deciding whether a period_prefix should be treated like a MemberRef postfix production instead of a new UnresolvedMemberExpr, and share it with the generic type disambiguation code, so that 'Foo<T>\n.bar()' parses consistently with 'Foo\n.bar()'. Extend the condition to consider a '.foo' reference on a line by itself in postfix position as a postfix continuation as well, allowing properties to be chained in builder fashion too. Addresses rdar://problem/20238557, but we may want to still make a more general change to the grammar here.

Swift SVN r28707