SILGen: Object to metatype casts have to take the slow path
It looks like we don't know how to open-code the "is-a metatype" check
for scalar casts, so just tighten up canUseScalarCheckedCastInstructions()
so that if the target type is a metatype, the source also has to be a
metatype.

This fixes a regression from "SIL: Use scalar casts more".

Fixes <rdar://problem/21003923>.

Swift SVN r28729